Refactor to avoid late imports
