+++ pyunifiprotect/data/nvr.py
@@ -3,7 +3,6 @@ from __future__ import annotations
 
 from datetime import datetime, timedelta, tzinfo
 from ipaddress import IPv4Address
-import logging
 from pathlib import Path
 from typing import Any, Dict, List, Literal, Optional, Set
 from uuid import UUID
@@ -16,20 +15,15 @@ from pyunifiprotect.data.base import (
     ProtectDeviceModel,
     ProtectModel,
     ProtectModelWithId,
-    create_from_unifi_dict,
 )
-from pyunifiprotect.data.devices import Bridge, Camera, Light, Sensor, Viewer
+from pyunifiprotect.data.devices import Camera
 from pyunifiprotect.data.types import (
     DoorbellMessageType,
     EventType,
-    FixSizeOrderedDict,
     ModelType,
     SmartDetectObjectType,
 )
-from pyunifiprotect.data.websocket import WSJSONPacketFrame, WSPacket
 from pyunifiprotect.utils import process_datetime
-
-_LOGGER = logging.getLogger(__name__)
 
 MAX_SUPPORTED_CAMERAS = 256
 MAX_EVENT_HISTORY_IN_STATE_MACHINE = MAX_SUPPORTED_CAMERAS * 2
@@ -436,127 +430,3 @@ class Liveview(ProtectModelWithId):
         """
 
         return self.api.bootstrap.users.get(self.owner_id)
-
-
-class Bootstrap(ProtectBaseObject):
-    auth_user_id: str
-    access_key: str
-    cameras: Dict[str, Camera]
-    users: Dict[str, User]
-    groups: Dict[str, Group]
-    liveviews: Dict[str, Liveview]
-    nvr: NVR
-    viewers: Dict[str, Viewer]
-    lights: Dict[str, Light]
-    bridges: Dict[str, Bridge]
-    sensors: Dict[str, Sensor]
-    last_update_id: UUID
-
-    # TODO:
-    # legacyUFVs
-    # displays
-    # doorlocks
-    # chimes
-    # schedules
-
-    # not directly from Unifi
-    events: Dict[str, Event] = FixSizeOrderedDict()
-
-    @classmethod
-    def unifi_dict_to_dict(cls, data: Dict[str, Any]) -> Dict[str, Any]:
-        for model_type in ModelType.bootstrap_models():
-            key = model_type + "s"
-            items: Dict[str, ProtectModel] = {}
-            for item in data[key]:
-                items[item["id"]] = item
-            data[key] = items
-
-        return super().unifi_dict_to_dict(data)
-
-    def unifi_dict(self, data: Optional[Dict[str, Any]] = None, exclude: Optional[Set[str]] = None) -> Dict[str, Any]:
-        data = super().unifi_dict(data=data, exclude=exclude)
-
-        if "events" in data:
-            del data["events"]
-
-        for model_type in ModelType.bootstrap_models():
-            attr = model_type + "s"
-            if attr in data and isinstance(data[attr], dict):
-                data[attr] = list(data[attr].values())
-
-        return data
-
-    @property
-    def auth_user(self) -> User:
-        user: User = self.api.bootstrap.users[self.auth_user_id]
-        return user
-
-    def process_event(self, event: Event) -> None:
-        if event.camera is None:
-            return
-
-        if event.type == EventType.MOTION:
-            if event.end is None:
-                event.camera.is_motion_detected = True
-            else:
-                event.camera.is_motion_detected = False
-
-                event.camera.last_motion = event.end
-                event.camera.last_motion_event_id = event.id
-        elif event.type == EventType.SMART_DETECT:
-            if event.end is not None:
-                event.camera.last_smart_detect = event.end
-                event.camera.last_smart_detect_event_id = event.id
-        elif event.type == EventType.RING:
-            event.camera.last_ring = event.start
-            event.camera.last_ring_event_id = event.id
-
-        self.events[event.id] = event
-
-    def process_ws_packet(self, packet: WSPacket) -> None:
-        if not isinstance(packet.action_frame, WSJSONPacketFrame):
-            _LOGGER.debug("Unexpected action frame format: %s", packet.action_frame.payload_format)
-
-        if not isinstance(packet.data_frame, WSJSONPacketFrame):
-            _LOGGER.debug("Unexpected data frame format: %s", packet.data_frame.payload_format)
-
-        action: dict = packet.action_frame.data  # type: ignore
-        data: dict = packet.data_frame.data  # type: ignore
-        self.last_update_id = UUID(action["newUpdateId"])
-
-        if action["modelKey"] not in ModelType.values():
-            _LOGGER.debug("Unknown model type: %s", action["modelKey"])
-            return
-
-        if action["action"] == "add":
-            obj = create_from_unifi_dict(data, api=self._api)
-
-            if isinstance(obj, Event):
-                self.process_event(obj)
-            elif (
-                isinstance(obj, ProtectModelWithId)
-                and obj.model is not None
-                and obj.model.value in ModelType.bootstrap_models()
-            ):
-                key = obj.model.value + "s"
-                getattr(self, key)[obj.id] = obj
-            else:
-                _LOGGER.debug("Unexpected bootstrap model type for add: %s", obj.model)
-        elif action["action"] == "update":
-            model_type = action["modelKey"]
-            if model_type in ModelType.bootstrap_models() or model_type == ModelType.EVENT.value:
-                key = model_type + "s"
-                devices = getattr(self, key)
-                if action["id"] in devices:
-                    obj: ProtectModel = devices[action["id"]]
-                    obj = obj.update_from_unifi_dict(data)
-
-                    if isinstance(obj, Event):
-                        self.process_event(obj)
-
-                    devices[action["id"]] = obj
-                # ignore updates to events that phase out
-                elif model_type != ModelType.EVENT.value:
-                    _LOGGER.debug("Unexpected %s: %s", key, action["id"])
-            else:
-                _LOGGER.debug("Unexpected bootstrap model type for update: %s", model_type)
